requirements.txt
asgiref==3.2.3
autopep8==1.5
begins==0.9
Django==3.0.3
django-crispy-forms==1.8.1
django-extensions==2.2.7
docutils==0.16
Pillow==7.0.0
pip-autoremove==0.9.1
pip-upgrade-outdated==1.5
pycodestyle==2.5.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
